requirements.txt
customtkinter
passlib
PyYAML
Pillow
